Moved Jolt hooks to extension initialization/termination
