+++ src/jolt_hooks.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+void initialize_jolt_hooks();
+
+void deinitialize_jolt_hooks();
