+++ src/register_types.cpp
@@ -1,4 +1,5 @@
 #include "jolt_debug_geometry_3d.hpp"
+#include "jolt_hooks.hpp"
 #include "jolt_physics_direct_body_state_3d.hpp"
 #include "jolt_physics_direct_space_state_3d.hpp"
 #include "jolt_physics_server_3d.hpp"
@@ -13,6 +14,8 @@ void on_initialize(ModuleInitializationLevel p_level) {
 		case MODULE_INITIALIZATION_LEVEL_CORE: {
 		} break;
 		case MODULE_INITIALIZATION_LEVEL_SERVERS: {
+			initialize_jolt_hooks();
+
 			ClassDB::register_class<JoltPhysicsDirectBodyState3D>();
 			ClassDB::register_class<JoltPhysicsDirectSpaceState3D>();
 			ClassDB::register_class<JoltPhysicsServer3D>();
@@ -42,6 +45,8 @@ void on_terminate(ModuleInitializationLevel p_level) {
 				memdelete(server_factory);
 				server_factory = nullptr;
 			}
+
+			deinitialize_jolt_hooks();
 		} break;
 		case MODULE_INITIALIZATION_LEVEL_SCENE: {
 		} break;
